optional-requirements.txt
pyzmq>=2.2.0
pbkdf2>=1.3
MySQL-python>=1.2.3
torndb>=0.1
hiredis>=0.1.1
msgpack-python>=0.1.13
pycassa>=1.6.0
pymongo>=2.1
redis>=2.4.12
psycopg2>=2.4.5
pycrypto>=2.6.1
